backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis[asyncio]==5.1.1
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-multipart==0.0.12
aiofiles==24.1.0
